logging: use a custom logging interface
The pkg namespace should not use an external top level logger.
This can cause problems if multiple packages such as kind/pkg are vendored
in a third party project and each one of them uses a custom top level
logger.

Arguably top level loggers such as klog and logrus should only be exposed
on the cmd side.

Introduce a Logger interface in kind/pkg/log and implement an object
that can be used for logging in kind/cmd. Currently the object is of type
logrus.Logger.