haproxy: prepull the image during build
Currently the haproxy image is being pulled when
haproxy is being run on the LB nodes.

To prevent that, prepull the image and support air gapped
HA setups. This requires finding what tag was preloaded
for the haproxy image.

In case no haproxy tags are found in the node-image
pull the default tag.