Use tar to archive logs
`kind export logs` can flake on very rare occasions.
In oder to export all logs and follow the symlinks we are exporting
one file at a time. This seems to cause sporadic issues with
docker exec logging exit status 126 errors.

This patch exports the /var/log folder of the nodes using tar
to reduce the number of streams needed.