Don't use proxy for the local network
On environments that require proxies to work, is a common issue
to forget setting the NO_PROXY environment variable to avoid the
kind nodes to try to use the proxy for the internal communications.
This patch automatically adds the docker network and the localhost
addresses to the NO_PROXY variable in case the environment needs
a proxy.