enables relative path for extraMounts
This enables a user to define a relative path for volumes and files that
are specified in extraMounts.

with this change a configuration like the following will determine the
path to adv-audit.yaml as relative to where `kind` is run. If you
specify a fully qualified path that will also be honored.
```
kind: Cluster
apiVersion: kind.sigs.k8s.io/v1alpha3
networking:
nodes:
- role: control-plane
  extraMounts:
  - containerPath: /etc/kubernetes/policies/adv-audit.yaml
    hostPath: adv-audit.yaml
    readOnly: true
    type: File
- role: worker
```

if you specify a hostPath that doesn't exist you get an error like:
```
Creating cluster "kind" ...
invalid configuration for node 0: Invalid hostPath: hostPath doesn't exist!: /home/dcooley/git/kind-k8scop/advaudit.yaml
ERROR: aborting due to invalid configuration

```

Use Case:

When you have files that are checked in as part of a github repository
it's helpful to use relative paths so that the kind config will work
regardless of where the files are checked out :)

Signed-off-by: Duffie Cooley <cooleyd@vmware.com>