Docs for using Ambassador as Ingress Controller
This commit adds support for installing Ambassador
(https://getambassador.io/) as an Ingress Controller.

Ambassador is installed via the Ambassador Operator which
is the recommended way of installed Ambassador (it makes sure
users always have the latest version of Ambassador installed
and takes care of the update schedule as well).