implement dns caching
since we implemented custom docker bridges, we use hostnames
instead of ip addresses on all the components.
This allows us to not care about the nodes IP addresses, so there
is no problem if the cluster is rebooted and other IPs are assigned
to the node.
However, this is not free, and it comes at the cost that every
hostname resolution is a DNS request, and by default uses UDP.
On noise environments this can cause packet drops, the most common
solution to avoid this is implement caching.
Linux has implemented caching in the glibc library using the nscd
daemon, that seems a good fit for us because does not require to
change the /etc/resolv.conf, instead other alternatives like dnsmasq.