use iptables-nft to avoid xtables.lock
iptables-nft and iptables-legacy are compatible, however, kubernetes
needs that all the components: kubelet and kube-proxy agree on the
userspace tools.

In addition, docker insert rules in the container to do some "magic"
with the DNS that we modify in order to survive reboots.

We use nft for kubernetes and pick the same version used by the
host to configure the magic DNS rules.