use /16 service subnet by default
the service subnet use a bitmap stored in etcd. Using a /12 subnet
imply that this bitmap can use 2^20 bits, something unnecessary
since Kubernetes is not validating more than 10K services.
A /16 subnet allows until 65536 services, more than enough, and
from the networking perspective is more correct, because it belong
to the Class B subnet and allows a better subnet planning.