Add LICENSE to site examples
Adding an explicit license to the example directory clarifies the
licensing of code snippets that have been included in the
documentation.

The license of these examples was previously up to interpretation as
the presence of a LICENSE file in `site/` (cf. [0]) implied CC-BY-SA
4.0 even though licensing code under that license is recommended
against even by creative commons themselves (cf. [1]).

Including this LICENSE file makes it clear that examples are licensed
under ASLv2 which makes it easier to adapt and integrate their code
into other projects.

[0]: https://github.com/kubernetes-sigs/kind/blob/master/site/LICENSE
[1]: https://creativecommons.org/faq/#can-i-apply-a-creative-commons-license-to-software