default ipv6 subnet to /56
we were defaulting IPv6 subnet to a /64, this was working because
kubeadm was calculating and modifying directly the node-cidr-mask,
so the controller manager does not complain.

Kubeadm should only validate that the config is correct and fail
in case of error, this behavior will be changed soon, but once
this happen it will break kind default deployment.