kindnetd don´t wait for kube-proxy
kindnetd allows to replace the incluster configuration, with the
internal API server address, with an environment variable.
It is possible that the environment variable we configure changes
after a node reboots.

If the connectivity to the new address fails, it falls back to the
incluster configuration.

This removes the dependency on kube-proxy, because it needs to
install the corresponding rules to forward traffic to
the apiserver Service.