compress exported logs
kind export logs uses tar to pipe all the /var/log folder and untar
them in the host.

docker always work, but podman flakes exporting the logs because of:
unexpected EOF: error writing to ...

it seems podman uses conmon to the pipe operations, to avoid that
any issues due to conmon not handling correctly some characters, we
can compress the bytes stream.