[WIP] Support remote podman
Gather info about the provider using `podman info` instead of
looking on the host directly. This allows starting a cluster
when the host is remote, for instance running in a VM on
a MacOS or Windows host.

Not all the info available via `docker info` is available in
`podman info`, so logging a warning when using podman.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>