Skip GitHub Actions on doc only changes
We have a set of GitHub Action jobs that run to test cgroupv2, docker,
and podman functionality. They currently run on all PRs to main.

These are functional tests, so they are not necessary when making
documentation changes. This updates the job definitions to skip running
if the PR only contains doc changes. This is likely rare, but will save
some resources by not testing when there are no code changes.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>