Add node build steps to Building Images section
In our quick start guide we have a Building Images section that talks
about being able to build custom images. It describes some of the
pieces, but it doesn't actually have the steps to take to actually
create the image.

There are notes in the repo README.md file that include these steps.
This adds a small amount of duplication, but this adds those steps to
the Building Images section to make it easier for folks to find out what
steps to take to build their images.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>