Adding support for 'kind provider' command
There are use cases where it would help to be able
to list the supported providers as well as display
the currently configured runtime provider.

This change adds a 'kind provider' command which
will display all supported providers. The --current
flag can be provided to show the currently configured
provider which honors the KIND_EXPERIMENTAL_PROVIDER
env if set.

Signed-off-by: Yibo Zhuang <yibzhuang@gmail.com>