avoid mismatches problems between iptables versions
As explained in https://bugzilla.netfilter.org/show_bug.cgi?id=1632
there is no compatibility guarantee for the userspace iptables
binaries, this causes that the output of iptables-save may be
different if the iptables host version doesn't match the version used
inside the kind node.

Since the iptables we need to process for doing the dns magic are well
known and can not change without breaking a lot of users, instead of
manipulating the existing rules, just obtain the parameters that we
need and remove and add the necessary rules.