Add third party package installation note
This restructures the installation instructions for kind to have the
first party, community supported methods listed first in the docs.

Third party maintained package installation instructions are moved to
the end of the list of installation options, and a note of warning is
added to point out that these packages are not maintained by the project
maintainers.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>