Mention possible pids_limit issues with rootless podman
When running rootless podman, there can be issues with processes not
able to create new pids. This is caused by podman's default limit being
too low for scenarios like running nginx and spawning workers.

This simply adds a notice to the rootless section and suggests a way to
disable said limit if desired.