Retry obtaining KUBECONFIG file lock
When there are multiple concurrent create or delete operations, there
can be different processing trying to lock the KUBECONFIG file at the
same time. This leads to small windows where getting the lock files.

We don't fail the operation if something like "kind delete cluster"
isn't able to remove its settings from KUBECONFIG. This is ideal and
what most users would expect, but it also means there may be some
leftover crud left behind that they are not aware of.

This adds some basic retries to obtaining the file lock. It potentially
adds a very small delay (in a much longer process), but makes it more
likely that the process is able to get the lock and complete the
KUBECONFIG update.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>