node build: Extract version from remote URL
At some point changes were made to the structure of the community
published tarballs where the expected "version" file is no longer
present. This causes the build to fail due to not being able to capture
the version of the source being built.

Most of the time the requested version is part of the URL for the remote
source anyway. This creates a fallback where the version will be
extracted from the URL if it cannot be read explicitly from a "version"
file in the tarball content.

This does not cover the local tarball file path, or if the user points
to a non-community published URL. But in the "normal" usage of pointing
to community published tarballs, this will prevent the node build from
failing in the cases where there is not a version file present.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>