Update golangci-lint to v2.3.1
Golangci-lint has released a new major version that includes some
changes to how it is configured, along with the usual linter updates
that come with every release.

This updates the golangci-lint version we use to v2.3.1 (the current
latest release) and updates the config file to conform to the new
format. Also includes quite a few trivial updates to address the new
warnings from the updated linters.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>