loadbalancer/config: Parse resolv.conf for DNS
Control plane in HA setup doesn't currently work using Podman.
This is because the haproxy DNS server is hard-coded to use
127.0.0.1:53, which is not available on Podman.

To get the actual NS address, /etc/resolv.conf must be consulted.
So this patch adds parse-resolv-conf to the haproxy resolvers block,
which makes HA control plane work on Podman instantly.

127.0.0.1:53 is retained as a nameserver still.